% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3.R, R/redis.R
\name{completeTask}
\alias{completeTask}
\alias{completeTask.RedisTQ}
\title{Lightweight Reliable Task Queue - Stop processing a task item}
\usage{
completeTask(tq, item)

\method{completeTask}{RedisTQ}(tq, item)
}
\arguments{
\item{tq}{Task Queue}

\item{item}{item obtained from \code{\link{leaseTask}}}
}
\description{
Lightweight Reliable Task Queue - Stop processing a task item
}
\seealso{
\code{\link{leaseTask}} \code{\link{completeTask}} \code{\link{createTask}}

\code{\link{RedisTQ}}
}
