% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3.R, R/redis.R
\name{createTask}
\alias{createTask}
\alias{createTask.RedisTQ}
\title{Lightweight Reliable Task Queue - Add a task item}
\usage{
createTask(tq, item)

\method{createTask}{RedisTQ}(tq, item)
}
\arguments{
\item{tq}{Task Queue}

\item{item}{task item to add}
}
\description{
Lightweight Reliable Task Queue - Add a task item
}
\seealso{
\code{\link{leaseTask}} \code{\link{completeTask}} \code{\link{createTask}}

\code{\link{RedisTQ}}
}
