% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3.R, R/redis.R
\name{existsTask}
\alias{existsTask}
\alias{existsTask.RedisTQ}
\title{Lightweight Reliable Task Queue - Check for task item availability}
\usage{
existsTask(tq)

\method{existsTask}{RedisTQ}(tq)
}
\arguments{
\item{tq}{Task Queue}
}
\description{
Lightweight Reliable Task Queue - Check for task item availability
}
\seealso{
\code{\link{leaseTask}} \code{\link{completeTask}} \code{\link{createTask}}

\code{\link{RedisTQ}}
}
